# GoMailFlows/docs/openapi.yaml
openapi: 3.1.0
info:
  title: HiTechCloud MailFlows API
  version: 1.0.0
  description: |
    HiTechCloud MailFlows - Enterprise Email SaaS Platform API
    
    ## Features
    - **Domain Management**: Register and manage email domains with automatic DNS provisioning
    - **Email Accounts**: Create and manage email mailboxes
    - **SMTP Configuration**: Per-domain or global SMTP relay configuration
    - **JMAP Support**: Full JMAP protocol (RFC 8620/8621) for modern email access
    - **Email Marketing**: Campaign management, subscriber lists, analytics
    - **SaaS Packages**: Flexible service plans with granular feature control
    - **Admin Panel**: Complete system administration capabilities
    - **DNS Templates**: Customizable DNS record templates for domain provisioning
    - **MCP Integration**: Model Context Protocol support for AI-powered management
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Obtain tokens via `/api/v1/auth/login` or `/api/v1/auth/register`.
    
    ## Rate Limiting
    API endpoints are rate-limited. Limits vary by endpoint and user package.
  contact:
    name: HiTechCloud Support
    url: https://hitechcloud.io/support
    email: support@hitechcloud.io
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.mailflows.hitechcloud.io/api/v1
...
security:
  - BearerAuth: []
paths:
...
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: User list
...
  /admin/dns-templates:
    get:
      tags: [admin]
      summary: List DNS templates
      responses:
        '200':
          description: DNS template list

    post:
      tags: [admin]
      summary: Create DNS template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDNSTemplateRequest'
      responses:
        '201':
          description: Template created
  /admin/settings:
    get:
      tags: [admin]
      summary: Get system settings
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Settings list
  /admin/audit-logs:
    get:
      tags: [admin]
      summary: Get audit logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Audit log list

# ============================================================
# Webhooks
# ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/register
  schemas:
...
    CreateDNSTemplateRequest:
      type: object
      required: [name, records]
      properties:
        name:
          type: string
        description:
          type: string
        is_default:
          type: boolean
        records:
          type: array
          minItems: 1
          items:
            type: object
            required: [type, host, value]
            properties:
              type:
                type: string
                enum: [MX, A, AAAA, CNAME, TXT, SRV, CAA, NS]
              host:
                type: string
                description: "Supports {{domain}} placeholder"
              value:
                type: string
                description: "Supports {{domain}}, {{server_ip}}, {{dkim_public_key}} placeholders"
              priority:
                type: integer
                nullable: true
              ttl:
                type: integer
                default: 3600
...